refactor: Answer 패키지 리팩터링
