+++ src/main/java/com/example/gistcompetitioncnserver/answer/application/PostDeleteEventListenerOfAnswer.java
@@ -1,5 +1,6 @@
-package com.example.gistcompetitioncnserver.answer;
+package com.example.gistcompetitioncnserver.answer.application;
 
+import com.example.gistcompetitioncnserver.answer.domain.AnswerRepository;
 import com.example.gistcompetitioncnserver.post.PostDeleteEvent;
 import org.springframework.context.ApplicationListener;
 import org.springframework.stereotype.Component;
