+++ src/main/java/com/example/gistcompetitioncnserver/answer/dto/AnswerRequest.java
@@ -1,4 +1,4 @@
-package com.example.gistcompetitioncnserver.answer;
+package com.example.gistcompetitioncnserver.answer.dto;
 
 import javax.validation.constraints.NotBlank;
