+++ src/main/java/com/example/gistcompetitioncnserver/answer/presentation/AnswerController.java
@@ -1,5 +1,8 @@
-package com.example.gistcompetitioncnserver.answer;
+package com.example.gistcompetitioncnserver.answer.presentation;
 
+import com.example.gistcompetitioncnserver.answer.dto.AnswerRequest;
+import com.example.gistcompetitioncnserver.answer.application.AnswerService;
+import com.example.gistcompetitioncnserver.answer.domain.Answer;
 import com.example.gistcompetitioncnserver.config.annotation.LoginUser;
 import com.example.gistcompetitioncnserver.config.annotation.ManagerPermissionRequired;
 import com.example.gistcompetitioncnserver.user.SimpleUser;
