+++ src/test/java/com/example/gistcompetitioncnserver/answer/AnswerServiceTest.java
@@ -1,6 +1,10 @@
 package com.example.gistcompetitioncnserver.answer;
 
 import com.example.gistcompetitioncnserver.ServiceTest;
+import com.example.gistcompetitioncnserver.answer.application.AnswerService;
+import com.example.gistcompetitioncnserver.answer.domain.Answer;
+import com.example.gistcompetitioncnserver.answer.domain.AnswerRepository;
+import com.example.gistcompetitioncnserver.answer.dto.AnswerRequest;
 import com.example.gistcompetitioncnserver.exception.WrappedException;
 import com.example.gistcompetitioncnserver.exception.post.NoSuchPostException;
 import com.example.gistcompetitioncnserver.exception.post.UnAnsweredPostException;
